reorganize slightly
