+++ client/src/index.jsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { createStore } from 'redux'
-import todoApp from './react/reducer'
+import todoApp from './components/reducer'
 
 import { AppContainer } from 'react-hot-loader';
 // AppContainer is a necessary wrapper component for HMR
 
-import App from './react/app.jsx';
+import App from './components/app.jsx';
 
 let store = createStore(todoApp);
 
@@ -23,7 +23,7 @@ render(App);
 
 // Hot Module Replacement API
 if (module.hot) {
-    module.hot.accept('./react/app.jsx', () => {
+    module.hot.accept('./components/app.jsx', () => {
         render(App)
     });
 }
